@@ -51,7 +51,7 @@
     "shelljs": "0.6.0",
     "tsd": "0.6.0-beta.5",
     "typedoc": "0.3.12",
-    "typescript": "2.0.3",
+    "typescript": "2.1.5",
     "webfont": "^4.0.0"
   },
   "jshintConfig": {
