@@ -50,6 +50,12 @@ const COMMAND_TYPE_AND_CR_SELECTION = "typeSelectionAndCr";
 const COMMAND_TYPE_SELECTION = "typeSelection";
 const COMMAND_OPEN_COMMAND_PALETTE = CommandPaletteRequestTypes.COMMAND_OPEN_COMMAND_PALETTE;
 
+const COMMAND_SYNTAX_HIGHLIGHTING = "syntaxHighlighting";
+const EVENT_COMMAND_SYNTAX_HIGHLIGHTING = "TEXTVIEWER_EVENT_COMMAND_SYNTAX_HIGHLIGHTING";
+
+const COMMAND_TAB_SIZE = "tabSize";
+const EVENT_COMMAND_TAB_SIZE = "TEXTVIEWER_EVENT_COMMAND_TAB_WIDTH";
+
 const COMMANDS = [
   COMMAND_TYPE_AND_CR_SELECTION,
   COMMAND_TYPE_SELECTION,
@@ -144,7 +150,7 @@ class EtTextViewer extends ViewerElement implements CommandPaletteRequestTypes.C
     this._mimeType = null;
     this._log = new Logger(EtTextViewer.TAG_NAME, this);
     this._commandLine = null;
-    this._returnCode  =null;
+    this._returnCode = null;
     this._editable = false;
     this._codeMirror = null;
     this._height = 0;
@@ -256,8 +262,10 @@ class EtTextViewer extends ViewerElement implements CommandPaletteRequestTypes.C
     this._mimeType = mimeType;
     
     const modeInfo = CodeMirror.findModeByMIME(mimeType);
-    if (modeInfo.mode !== undefined && modeInfo.mode !== null && modeInfo.mode !== "null") {
-      LoadCodeMirrorMode(modeInfo.mode);
+    if (modeInfo != null) {
+      if (modeInfo.mode !== null && modeInfo.mode !== "null") {
+        LoadCodeMirrorMode(modeInfo.mode);
+      }
       if (this._codeMirror !== null) {
         this._codeMirror.setOption("mode", mimeType);
       }
@@ -267,7 +275,15 @@ class EtTextViewer extends ViewerElement implements CommandPaletteRequestTypes.C
   get mimeType(): string {
     return this._mimeType;
   }
-  
+
+  getTabSize(): number {
+    return parseInt(this._codeMirror.getOption("tabSize"), 10);
+  }
+
+  setTabSize(size: number): void {
+    this._codeMirror.setOption("tabSize", size);
+  }
+
   setBytes(buffer: Buffer, mimeType: string): void {
     let charset = "utf-8";
     let cleanMimeType = mimeType;
@@ -753,6 +769,15 @@ class EtTextViewer extends ViewerElement implements CommandPaletteRequestTypes.C
     this.dispatchEvent(event);
   }
 
+  private _getMimeTypeName(): string {
+    for (const info of CodeMirror.modeInfo) {
+      if (info.mime === this._mimeType) {
+        return info.name;
+      }
+    }
+    return this._mimeType;
+  }
+
   scrollTo(optionsOrX: ScrollToOptions | number, y?: number): void {
     let xCoord = 0;
     let yCoord = 0;
@@ -886,13 +911,15 @@ class EtTextViewer extends ViewerElement implements CommandPaletteRequestTypes.C
     if (this._mode === ViewerElementTypes.Mode.DEFAULT) {
       ev.stopPropagation();
       ev.preventDefault();
-    }      
+    }
   }
   
   private _commandPaletteEntries(): CommandPaletteRequestTypes.CommandEntry[] {
     let commandList: CommandPaletteRequestTypes.CommandEntry[] = [
       { id: COMMAND_TYPE_SELECTION, group: PALETTE_GROUP, iconRight: "terminal", label: "Type Selection", target: this },
-      { id: COMMAND_TYPE_AND_CR_SELECTION, group: PALETTE_GROUP, iconRight: "terminal", label: "Type Selection & Execute", target: this }
+      { id: COMMAND_TYPE_AND_CR_SELECTION, group: PALETTE_GROUP, iconRight: "terminal", label: "Type Selection & Execute", target: this },
+      { id: COMMAND_SYNTAX_HIGHLIGHTING, group: PALETTE_GROUP, iconRight: "", label: "Syntax: " + this._getMimeTypeName(), target: this },
+      { id: COMMAND_TAB_SIZE, group: PALETTE_GROUP, iconRight: "", label: "Tab Size: " + this.getTabSize(), target: this }
     ];
     
     if (this._mode ===ViewerElementTypes.Mode.CURSOR) {
@@ -956,7 +983,15 @@ class EtTextViewer extends ViewerElement implements CommandPaletteRequestTypes.C
           commandPaletteRequestDetail);
         this.dispatchEvent(commandPaletteRequestEvent);
         break;
-        
+
+      case COMMAND_SYNTAX_HIGHLIGHTING:
+        this.dispatchEvent(new CustomEvent(EVENT_COMMAND_SYNTAX_HIGHLIGHTING, {bubbles: true, composed: true, detail: { srcElement: this } } ));
+        break;
+
+      case COMMAND_TAB_SIZE:
+        this.dispatchEvent(new CustomEvent(EVENT_COMMAND_TAB_SIZE, {bubbles: true, composed: true, detail: { srcElement: this } } ));
+        break;
+
       default:
         if (this._mode === ViewerElementTypes.Mode.CURSOR && CodeMirrorCommands.isCommand(command)) {
           CodeMirrorCommands.executeCommand(this._codeMirror, command);
